Refactorin Contacts and Membertable
